+++ .eslintrc.js
@@ -1,6 +1,13 @@
 module.exports = {
 	extends: '@chipzhang/eslint-config/node',
 	rules: {
-		'no-console': 'off',
+		'node/shebang': [
+			'error',
+			{
+				convertPath: {
+					'**/*.ts': ['^(.*)\\.ts$', '$1.js'],
+				},
+			},
+		],
 	},
 }