+++ gen-types.d.ts
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+/**
+ * Generate `*.less.d.ts` for `*.less` file.
+ */
+export {};