+++ CHANGELOG.md
@@ -1,12 +1,16 @@
 # Change Log
 
+## 2021-05-24 - 0.1.3
+
+Convert `index.js`, `parse.js`, `gen-types.js` to TypeScript
+
 ## 2021-05-22 - 0.1.2
 
 Fix an incorrectness in `README.md`
 
 ## 2021-05-22 - 0.1.1
 
-Added type definitions `utils.d.ts`
+Convert `utils.js` to TypeScript
 
 ## 2021-05-20 - 0.1.0